Preserve old directories on save.
When the admin clicks save, and the tree is not
fully expandable such that collection directories
are not checked, but their parents are undetermined,
we check if any existing collection path starts with
an undetermined parent. If it does we still keep it saved.

But after that, if it is specifically found unchecked,
we remove it.